Moved some forallN constructs and collapse-nowait execution policy to higher-level pattern, policy headers where they will be easier to see for uses who need to use them.
