+++ include/RAJA/internal/ForallNPolicy.hpp
@@ -69,24 +69,12 @@ struct ForallN_PolicyPair : public I {
   explicit constexpr ForallN_PolicyPair(ISET const &i) : ISET(i) {}
 };
 
-template <typename... PLIST>
-struct ExecList {
-  constexpr const static size_t num_loops = sizeof...(PLIST);
-  typedef std::tuple<PLIST...> tuple;
-};
-
 // Execute (Termination default)
 struct ForallN_Execute_Tag {
 };
 
 struct Execute {
   typedef ForallN_Execute_Tag PolicyTag;
-};
-
-template <typename EXEC, typename NEXT = Execute>
-struct NestedPolicy {
-  typedef NEXT NextPolicy;
-  typedef EXEC ExecPolicies;
 };
