+++ include/RAJA/pattern/forallN.hpp
@@ -66,6 +66,31 @@
 namespace RAJA
 {
 
+/*!
+ * \brief Struct used to define forallN nested policies.
+ *
+ *  Typically, passed as first template argument to forallN templates.
+ */
+template <typename EXEC, typename NEXT = Execute>
+struct NestedPolicy {
+  typedef NEXT NextPolicy;
+  typedef EXEC ExecPolicies;
+};
+
+/*!
+ * \brief Struct that contains a policy for each loop nest in a forallN
+ *        construct. 
+ *
+ *  Typically, passed as first template argument to NestedPolicy template,
+ *  followed by permutation, etc.
+ */
+template <typename... PLIST>
+struct ExecList {
+  constexpr const static size_t num_loops = sizeof...(PLIST);
+  typedef std::tuple<PLIST...> tuple;
+};
+
+
 /******************************************************************
  *  ForallN_Executor(): Default Executor for loops
  ******************************************************************/
@@ -75,7 +100,6 @@ namespace RAJA
  *
  *  The default action is to call RAJA::forall to peel off outer loop nest.
  */
-
 template <typename POLICY_INIT, typename... POLICY_REST>
 struct ForallN_Executor<POLICY_INIT, POLICY_REST...> {
   typedef typename POLICY_INIT::ISET TYPE_I;
